pyproject.toml
[project]
name = "notify"
version = "0.1.0"
description = "Internal multi-channel notification service (WhatsApp, SMS, Email)"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlmodel>=0.0.22",
    "alembic>=1.14",
    "pydantic-settings>=2.6",
    "python-multipart>=0.0.12",
    "jinja2>=3.1",
    "httpx>=0.27",
    "aiosmtplib>=3.0",
    "markdown-it-py>=3.0",
    "linkify-it-py>=2.0",
    "rq>=1.16",
    "redis>=5.2",
    "qrcode[pil]>=8.0",
    "email-validator>=2.2",
    "typer[all]>=0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "httpx>=0.27",
]

[project.scripts]
notify = "app.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
